Fix monospaced font in log dialog
When using Qt >= 5.2, query QFontDatabase for the default monospaced
font, which will depend on the actual platform.
On previous versions (including Qt 4), build a font using only the
"Monospace" name, with the right hint for requesting something
"monospaced" -- even if the platform does not have a font named that
way (e.g. on Windows), Qt's font matching algo will try to find a font
fitting the typewriter request.

Also, do not hardcode a size, as either what QFontDatabase returns
or the QFont constructor will provide the default size on the platform.